refactor: 拆分 models 和 handlers 避免循环导入
