+++ src/plugins/github/handlers/__init__.py
@@ -1,0 +1,3 @@
+from .git import GitHandler as GitHandler
+from .github import GithubHandler as GithubHandler
+from .issue import IssueHandler as IssueHandler
