+++ tests/plugins/github/publish/utils/test_comment_issue.py
@@ -5,7 +5,8 @@ from tests.plugins.github.utils import get_github_bot
 
 
 async def test_comment_issue(app: App, mocker: MockerFixture):
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
 
     render_comment = "test"
 
@@ -40,7 +41,8 @@ async def test_comment_issue(app: App, mocker: MockerFixture):
 
 async def test_comment_issue_reuse(app: App, mocker: MockerFixture):
     from src.plugins.github.constants import NONEFLOW_MARKER
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
 
     render_comment = "test"
 
@@ -76,7 +78,8 @@ async def test_comment_issue_reuse(app: App, mocker: MockerFixture):
 
 async def test_comment_issue_reuse_same(app: App, mocker: MockerFixture):
     """测试评论内容相同时不会更新评论"""
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
 
     render_comment = "test\n<!-- NONEFLOW -->\n"
