+++ tests/plugins/github/remove/utils/test_remove_resolve_pull_requests.py
@@ -37,7 +37,8 @@ async def test_resolve_conflict_pull_requests_bot(
     app: App, mocker: MockerFixture, mocked_api: MockRouter, tmp_path: Path, mock_pull
 ) -> None:
     from src.plugins.github import plugin_config
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
     from src.plugins.github.plugins.remove.utils import resolve_conflict_pull_requests
     from src.providers.utils import dump_json5
 
@@ -140,7 +141,8 @@ async def test_resolve_conflict_pull_requests_plugin(
     app: App, mocker: MockerFixture, mocked_api: MockRouter, tmp_path: Path, mock_pull
 ) -> None:
     from src.plugins.github import plugin_config
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
     from src.plugins.github.plugins.remove.utils import resolve_conflict_pull_requests
     from src.providers.utils import dump_json5
 
@@ -247,7 +249,8 @@ async def test_resolve_conflict_pull_requests_not_found(
     app: App, mocker: MockerFixture, mocked_api: MockRouter, tmp_path: Path, mock_pull
 ) -> None:
     from src.plugins.github import plugin_config
-    from src.plugins.github.models import GithubHandler, RepoInfo
+    from src.plugins.github.handlers import GithubHandler
+    from src.plugins.github.models import RepoInfo
     from src.plugins.github.plugins.remove.utils import resolve_conflict_pull_requests
     from src.providers.utils import dump_json5
